fix windows build
34a7b51ea482d4487d4fb8e61cde18aa1ff9b8ff (for 5-1) auto → size_t made
32bit happy, but win64bit still failed with:
error C2784: ´_Ty std::min(std::initializer_list<_Elem>,_Pr)´ : could
not deduce template argument for ´std::initializer_list<_Elem>´ from
´__int64´

(cherry picked from commit 53f645a9c959d93bde9230862c415c4ab2e3817b)

+++ vcl/win/source/gdi/winlayout.cxx
@@ -3907,8 +3907,8 @@ bool D2DWriteTextOutRenderer::DrawGlyphs(const Point & origin, uint16_t * pGid, 
         DWRITE_GLYPH_OFFSET glyphOffsets[MAX_GLYPHS] = { { 0.0f, 0.0f }, };
 
         mpRT->BeginDraw();
-        for (size_t nGlyphs = std::min(pGidEnd - pGid, MAX_GLYPHS); pGid != pGidEnd;
-                    pGid += nGlyphs, nGlyphs = std::min(pGidEnd - pGid, MAX_GLYPHS))
+        for (size_t nGlyphs = std::min(static_cast<int>(pGidEnd - pGid), MAX_GLYPHS); pGid != pGidEnd;
+                    pGid += nGlyphs, nGlyphs = std::min(static_cast<int>(pGidEnd - pGid), MAX_GLYPHS))
         {
             std::copy(pGid, pGidEnd, glyphIndices);
             std::copy_n(pAdvances, nGlyphs, glyphAdvances);
